Tweak logger to always log doubles for numbers, since we cannot reliably tell if a number will always be an int

+++ src/main/java/frc/lib/logging/JsonNodeLogger.java
@@ -81,7 +81,7 @@ public class JsonNodeLogger extends ClassSpecificLogger<JsonNode> {
         if(t == JsonToken.VALUE_NUMBER_FLOAT)
             backend.log(key, n.asDouble());
         else if(t == JsonToken.VALUE_NUMBER_INT)
-            backend.log(key, n.asInt());
+            backend.log(key, n.asDouble());
         else if(t == JsonToken.VALUE_TRUE || t == JsonToken.VALUE_FALSE)
             backend.log(key, n.asBoolean());
         else
